+++ src/zope/sendmail/interfaces.py
@@ -176,6 +176,13 @@ class ISMTPMailer(IMailer):
         title=_("Force TLS"),
         description=_("Use TLS always for sending email."))
 
+    implicit_tls = Bool(
+        title=_("Implicit TLS"),
+        description=_(
+            "Use TLS from the beginning of the connection, "
+            "known as SMTPS and commonly used on TCP port 465. "
+            "force_tls and no_tls are ignored if this is set."),)
+
 
 class IMaildirFactory(Interface):